-Adjusted YabauseGetTicks and fps code so it allows for greater precision -Changed windows code so it uses performance counters from now on.
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@1238 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/vdp2.h
@@ -358,7 +358,7 @@ typedef struct {
 } Vdp2Internal_struct;
 
 extern Vdp2Internal_struct Vdp2Internal;
-extern u32 lastticks;
+extern u64 lastticks;
 
 int Vdp2Init(void);
 void Vdp2DeInit(void);
